+++ app/jobs/accept-join-requests.js
@@ -1,7 +1,7 @@
 'use strict'
 const roblox = require('noblox.js')
 
-const DiscordMessageJob = require('../jobs/discord-message-job')
+const DiscordMessageJob = require('./discord-message')
 
 const trelloController = require('../services/trello')
