+++ app/services/group.js
@@ -8,7 +8,7 @@ const trelloService = require('./trello')
 
 const timeHelper = require('../helpers/time')
 
-const DiscordMessageJob = require('../jobs/discord-message-job')
+const DiscordMessageJob = require('../jobs/discord-message')
 
 exports.suspend = async (groupId, userId, options) => {
     const rank = await exports.getRank(groupId, userId)
